+++ kitsune/Cargo.toml
@@ -107,7 +107,7 @@ trials = { path = "../lib/trials" }
 typed-builder = "0.18.2"
 url = "2.5.0"
 utoipa = { version = "4.2.3", features = ["axum_extras", "uuid"] }
-utoipa-swagger-ui = { version = "6.0.0", features = ["axum"] }
+utoipa-swagger-ui = { version = "=6.0.0", features = ["axum"] }
 
 # --- Optional dependencies ---
